# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Online Examination Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # USERS
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # COURSES
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List courses
      parameters:
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # CLASSES
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List classes
      parameters:
        - name: courseId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /classes/{id}:
    patch:
      operationId: updateClass
      tags: [classes]
      summary: Update class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Delete class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # QUESTIONS
...
  /enrollments:
    get:
      operationId: listEnrollments
      tags: [enrollments]
      summary: List enrollments
      parameters:
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: courseId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
    post:
      operationId: createEnrollment
      tags: [enrollments]
      summary: Enroll student in course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentInput"
      responses:
        "201":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /enrollments/{id}:
    delete:
      operationId: deleteEnrollment
      tags: [enrollments]
      summary: Remove enrollment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ATTEMPTS
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # USER
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, admin, lecturer, student]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        studentId:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    Department:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        userCount:
          type: integer
        courseCount:
          type: integer
        createdAt:
          type: string
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
    DepartmentUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string

    # COURSE
    Course:
      type: object
      required: [id, name, code, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        description:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        lecturerId:
          type: ["integer", "null"]
        lecturerName:
          type: ["string", "null"]
        enrollmentCount:
          type: integer
        examCount:
          type: integer
        createdAt:
          type: string
    CourseInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
        departmentId:
          type: ["integer", "null"]
        lecturerId:
          type: ["integer", "null"]
    CourseUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
        departmentId:
          type: ["integer", "null"]
        lecturerId:
          type: ["integer", "null"]

    # CLASS
    Class:
      type: object
      required: [id, name, courseId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        courseId:
          type: integer
        courseName:
          type: ["string", "null"]
        academicYear:
          type: ["string", "null"]
        semester:
          type: ["string", "null"]
        studentCount:
          type: integer
        createdAt:
          type: string
    ClassInput:
      type: object
      required: [name, courseId]
      properties:
        name:
          type: string
        courseId:
          type: integer
        academicYear:
          type: string
        semester:
          type: string
    ClassUpdate:
      type: object
      properties:
        name:
          type: string
        academicYear:
          type: string
        semester:
          type: string

    # QUESTION
...
    Enrollment:
      type: object
      required: [id, studentId, courseId, enrolledAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        courseId:
          type: integer
        courseName:
          type: ["string", "null"]
        classId:
          type: ["integer", "null"]
        enrolledAt:
          type: string
    EnrollmentInput:
      type: object
      required: [studentId, courseId]
      properties:
        studentId:
          type: integer
        courseId:
          type: integer
        classId:
          type: ["integer", "null"]

    # ATTEMPT
...